+++ Source/CesiumRuntime/Public/CesiumFeatureIdAttribute.h
@@ -131,16 +131,16 @@ public:
       UPARAM(ref) const FCesiumFeatureIdAttribute& FeatureIDAttribute);
 
   /**
-   * Get the number of vertices in the primitive containing the feature
-   * ID attribute. If the feature ID attribute is invalid, this returns 0.
+   * Get the number of feature IDs in this attribute. If the feature ID
+   * attribute is invalid, this returns 0.
    */
   UFUNCTION(
       BlueprintCallable,
       BlueprintPure,
       Category = "Cesium|Features|FeatureIDAttribute")
   static int64
-  GetVertexCount(UPARAM(ref)
-                     const FCesiumFeatureIdAttribute& FeatureIDAttribute);
+  GetFeatureIDCount(UPARAM(ref)
+                        const FCesiumFeatureIdAttribute& FeatureIDAttribute);
 
   /**
    * Gets the feature ID at the given index. A feature ID can be used with a